# docs/swagger.yaml
basePath: /
definitions:
  github_com_ayMissouri_watchlist-go_git_internal_models.AnimeDetail:
    properties:
      description:
        type: string
      english_name:
        type: string
      episodes:
        description: 0 for unknown
        type: integer
      format:
        description: tv, movie, ova, ona, special, music
        type: string
      genres:
        items:
          type: string
        type: array
      id:
        description: MAL id
        type: string
      links:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.Link'
        type: array
      name:
        type: string
      poster:
        type: string
      released:
        type: string
      runtime:
        description: per episode
        type: string
      score:
        type: string
      status:
        description: finished_airing, currently_airing, not_yet_aired
        type: string
      studios:
        items:
          type: string
        type: array
      type:
        type: string
      year:
        type: string
    type: object
...
  github_com_ayMissouri_watchlist-go_git_internal_models.DiscoverAllResponse:
    properties:
      popular_movies:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.DiscoverItem'
        type: array
      popular_shows:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.DiscoverItem'
        type: array
      top_rated_movies:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.DiscoverItem'
        type: array
      top_rated_shows:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.DiscoverItem'
        type: array
    type: object
  github_com_ayMissouri_watchlist-go_git_internal_models.DiscoverItem:
    properties:
      background:
        type: string
      id:
        type: string
      imdb_rating:
        type: string
      poster:
        type: string
      title:
        type: string
      title_english:
        description: anime only
        type: string
      type:
        type: string
      year:
        type: string
    type: object
  github_com_ayMissouri_watchlist-go_git_internal_models.DiscoverResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.DiscoverItem'
        type: array
    type: object
  github_com_ayMissouri_watchlist-go_git_internal_models.Episode:
    properties:
      episode:
        type: integer
      filler:
        type: boolean
      id:
        type: string
      imdbRating:
        type: string
      overview:
        type: string
      recap:
        type: boolean
      released:
        type: string
      season:
        type: integer
      thumbnail:
        type: string
      title:
        type: string
    type: object
...
  github_com_ayMissouri_watchlist-go_git_internal_models.EpisodesResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.Episode'
        type: array
    type: object
...
  github_com_ayMissouri_watchlist-go_git_internal_models.Link:
    properties:
      category:
        type: string
      name:
        type: string
      url:
        type: string
    type: object
...
  github_com_ayMissouri_watchlist-go_git_internal_models.MovieDetail:
    properties:
      awards:
        type: string
      background:
        type: string
      behaviorHints:
        $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.BehaviorHints'
      cast:
        items:
          type: string
        type: array
      country:
        type: string
      credits:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.Credit'
        type: array
      description:
        type: string
      director:
        items:
          type: string
        type: array
      dvdRelease:
        type: string
      genre:
        items:
          type: string
        type: array
      genres:
        items:
          type: string
        type: array
      id:
        type: string
      imdb_id:
        type: string
      imdbRating:
        type: string
      links:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.Link'
        type: array
      logo:
        type: string
      moviedb_id:
        type: integer
      name:
        type: string
      popularities:
        $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.Popularities'
      popularity:
        type: number
      poster:
        type: string
      releaseInfo:
        type: string
      released:
        type: string
      runtime:
        type: string
      slug:
        type: string
      trailerStreams:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.TrailerStream'
        type: array
      trailers:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.Trailer'
        type: array
      type:
        type: string
      videos:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.Episode'
        type: array
      writer:
        items:
          type: string
        type: array
      year:
        type: string
    type: object
...
  github_com_ayMissouri_watchlist-go_git_internal_models.Person:
    properties:
      biography:
        type: string
      birthday:
        type: string
      credits:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.PersonCredit'
        type: array
      deathday:
        type: string
      id:
        type: integer
      imdb_id:
        type: string
      known_for:
        type: string
      name:
        type: string
      photo:
        type: string
      place_of_birth:
        type: string
    type: object
...
  github_com_ayMissouri_watchlist-go_git_internal_models.SearchResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.DiscoverItem'
        type: array
      query:
        type: string
      type:
        description: leave empty for mixed results
        type: string
    type: object
  github_com_ayMissouri_watchlist-go_git_internal_models.SeriesDetail:
    properties:
      awards:
        type: string
      background:
        type: string
      behaviorHints:
        $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.BehaviorHints'
      cast:
        items:
          type: string
        type: array
      country:
        type: string
      credits:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.Credit'
        type: array
      description:
        type: string
      director:
        items:
          type: string
        type: array
      dvdRelease:
        type: string
      genre:
        items:
          type: string
        type: array
      genres:
        items:
          type: string
        type: array
      id:
        type: string
      imdb_id:
        type: string
      imdbRating:
        type: string
      links:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.Link'
        type: array
      logo:
        type: string
      moviedb_id:
        type: integer
      name:
        type: string
      popularities:
        $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.Popularities'
      popularity:
        type: number
      poster:
        type: string
      releaseInfo:
        type: string
      released:
        type: string
      runtime:
        type: string
      slug:
        type: string
      status:
        type: string
      trailerStreams:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.TrailerStream'
        type: array
      trailers:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.Trailer'
        type: array
      tvdb_id:
        type: integer
      type:
        type: string
      videos:
        items:
          $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.Episode'
        type: array
      writer:
        items:
          type: string
        type: array
      year:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: ayMissouri
    url: https://github.com/ayMissouri/watchlist-go.git
  description: Backend API for tracking movies and TV shows.
  license:
    name: MIT
  title: Watchlist API
  version: "1.0"
paths:
...
  /discover:
    get:
      description: |-
        Returns one catalog. Use `sort` (optionally with `genre`) for popular/top-rated, `year` for a single release year, or `provider` for a streaming service. If you send more than one, `provider` wins over `year`, which wins over `sort`. Cached for an hour.
        Anime comes from MyAnimeList and only takes `sort`, which also allows `airing` and `upcoming` there.
      parameters:
      - description: Media type
        enum:
        - movie
        - series
        - anime
        in: query
        name: type
        required: true
        type: string
      - description: Sort order (default popular)
        enum:
        - popular
        - top_rated
        - airing
        - upcoming
        in: query
        name: sort
        type: string
      - description: 'Genre filter. Movies: action, adventure, animation, comedy,
          crime, documentary, drama, family, fantasy, history, horror, mystery, romance,
          sci-fi, thriller, war, western. Series: action, adventure, animation, comedy,
          crime, documentary, drama, family, fantasy, mystery, sci-fi, war, western,
          reality-tv, talk-show'
        in: query
        name: genre
        type: string
      - description: Release year, e.g. 2025 (overrides sort)
        in: query
        name: year
        type: string
      - description: Streaming provider (overrides sort and year)
        enum:
        - netflix
        - hbomax
        - disney
        - prime
        - appletv
        in: query
        name: provider
        type: string
      - description: Anime title language (default romaji, used when MAL has no title
          in that language)
        enum:
        - romaji
        - english
        - native
        in: query
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.DiscoverResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Discover movies or shows
      tags:
      - discover
  /discover/all:
    get:
      description: Popular movies, popular shows, top-rated movies and top-rated shows,
        all in one call. Cached for an hour.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.DiscoverAllResponse'
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Discover all catalogs
      tags:
      - discover
...
  /meta/{type}/{id}/recommendations:
    get:
      description: What TMDB (or MyAnimeList, for anime) recommends alongside this
        title, in the same shape as a discover catalog. Cached for a day.
      parameters:
      - description: Media type
        enum:
        - movie
        - series
        - anime
        in: path
        name: type
        required: true
        type: string
      - description: TMDB id (e.g. 278), MAL id for anime. IMDb ids like tt0111161
          still work for movies and series.
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.DiscoverResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get recommendations for a title
      tags:
      - meta
  /meta/anime/{id}:
    get:
      description: Everything MyAnimeList has on an anime. There's no episode list,
        only the `episodes` count. Cached for a day.
      parameters:
      - description: MyAnimeList id (e.g. 5114)
        in: path
        name: id
        required: true
        type: string
      - description: Anime title language (default romaji, used when MAL has no title
          in that language)
        enum:
        - romaji
        - english
        - native
        in: query
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.AnimeDetail'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get anime details
      tags:
      - meta
  /meta/anime/{id}/episodes:
    get:
      description: |-
        The episode list for an anime, all seasons flattened into season 1, the way MyAnimeList counts them.
        Titles and air dates come from a self-hosted Jikan instance (JIKAN_URL). When that isn't configured
        or can't be reached, episodes come back numbered ("Episode 1") with no air date, so the count is
        still right. Cached for a day.
      parameters:
      - description: MyAnimeList id (e.g. 5114)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.EpisodesResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get anime episodes
      tags:
      - meta
  /meta/movie/{id}:
    get:
      description: Everything we know about a movie. Cached for a day.
      parameters:
      - description: TMDB id (e.g. 278). IMDb ids like tt0111161 still work.
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.MovieDetail'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get movie details
      tags:
      - meta
  /meta/person/{id}:
    get:
      description: Bio and the most popular acting credits of a cast member. Ids come
        from `credits` on a movie or series detail; each credit is keyed like a discover
        item so it links to a detail page. Cached for a day.
      parameters:
      - description: TMDB person id (e.g. 2524)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.Person'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get person details
      tags:
      - meta
  /meta/series/{id}:
    get:
      description: Everything we know about a series, episodes included (they're in
        `videos`). Cached for a day.
      parameters:
      - description: TMDB id (e.g. 1396). IMDb ids like tt0903747 still work.
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.SeriesDetail'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get series details
      tags:
      - meta
...
  /search:
    get:
      description: |-
        Search by title. Leave `type` out and you get movies and shows mixed together, newest first.
        Anime is only searched with `type=anime` (MyAnimeList needs at least 3 characters).
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - description: Filter by type
        enum:
        - movie
        - series
        - anime
        in: query
        name: type
        type: string
      - description: Anime title language (default romaji, used when MAL has no title
          in that language)
        enum:
        - romaji
        - english
        - native
        in: query
        name: title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ayMissouri_watchlist-go_git_internal_models.SearchResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Search movies and shows
      tags:
      - search
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"